feat(Provider): creating get all and get by id api methods

+++ api-marketplace/src/main/java/br/com/apimarketplace/service/ApiCategoryService.java
@@ -14,7 +14,7 @@ public class ApiCategoryService {
         this.apiCategoryRepository = apiCategoryRepository;
     }
 
-    public Integer createCategory(CreateApiCategory createApiCategory) {
+    public Long createCategory(CreateApiCategory createApiCategory) {
         var entity = new ApiCategory(createApiCategory.name());
 
         var categorySaved = apiCategoryRepository.save(entity);
